Improved report generation
Changed report generation so doesn't need to reassess beam (and hence works much faster). New method involves converting figure to json and saving in the session. Visting the download url takes the json and converts it back to a plotly figure before changing to an image for the report. Report can now also be downloaded as HMTL report.